Conflicts resolved:
	chromium/base/BUILD.gn
	chromium/base/allocator/partition_allocator/spin_lock.cc -> partition_lock.cc
	chromium/base/files/file.h
	chromium/base/files/file_posix.cc
	chromium/base/files/file_util_posix.cc
	chromium/base/i18n/time_formatting.cc
	chromium/base/memory/platform_shared_memory_region.h
	chromium/base/memory/shared_memory_mapping.cc
	chromium/base/message_loop/message_loop_current.cc -> chromium/base/task/current_thread.cc
	chromium/base/message_loop/message_loop_current.h -> chromium/base/task/current_thread.h
	chromium/base/metrics/field_trial.cc
	chromium/base/metrics/field_trial.h
	chromium/base/posix/unix_domain_socket.cc
	chromium/base/process/launch.h
	chromium/base/process/launch_posix.cc
	chromium/base/process/memory.cc
	chromium/base/process/process_metrics.h
	chromium/base/process/process_metrics_posix.cc
	chromium/base/synchronization/condition_variable_posix.cc
	chromium/base/threading/platform_thread_unittest.cc
	chromium/base/time/time_exploded_posix.cc
	chromium/base/trace_event/process_memory_dump.cc
	chromium/build/build_config.h
	chromium/build/config/BUILD.gn
	chromium/build/config/compiler/BUILD.gn
	chromium/build/config/crypto.gni
	chromium/build/config/linux/nss/BUILD.gn
	chromium/build/config/linux/pangocairo/pangocairo.gni
	chromium/build/config/ui.gni
	chromium/components/autofill/core/common/autofill_regexes.cc
	chromium/components/services/font/BUILD.gn
	chromium/content/browser/BUILD.gn
	chromium/content/browser/child_process_launcher_helper_posix.cc
	chromium/content/browser/scheduler/responsiveness/native_event_observer.cc
	chromium/content/renderer/BUILD.gn
	chromium/content/renderer/discardable_memory_utils.cc
	chromium/extensions/browser/api/messaging/message_service.cc
	chromium/gpu/command_buffer/service/feature_info.cc
	chromium/gpu/command_buffer/service/mailbox_manager_sync.cc
	chromium/gpu/command_buffer/service/texture_definition.cc
	chromium/gpu/ipc/common/surface_handle.h
	chromium/gpu/vulkan/vulkan_export.h -> <deleted>
	chromium/ipc/BUILD.gn
	chromium/ipc/ipc_message_utils.cc
	chromium/media/capture/video/fake_video_capture_device_factory.cc
	chromium/mojo/core/channel.cc
	chromium/mojo/core/channel_posix.cc
	chromium/mojo/core/core.cc
	chromium/mojo/core/node_controller.cc
	chromium/mojo/core/shared_buffer_dispatcher.cc
	chromium/mojo/public/cpp/platform/socket_utils_posix.cc
	chromium/mojo/public/cpp/system/platform_handle.cc
	chromium/net/cert/cert_verifier.cc
	chromium/net/cert/cert_verify_proc.cc
	chromium/net/cert/cert_verify_proc.h
	chromium/net/cert/cert_verify_proc_nss.cc -> <deleted>
	chromium/net/cert/cert_verify_proc_unittest.cc
	chromium/net/cert/test_root_certs_unittest.cc
	chromium/net/dns/dns_reloader.cc
	chromium/net/dns/dns_reloader.h
	chromium/net/dns/host_resolver_manager.cc
	chromium/net/dns/host_resolver_proc.cc
	chromium/net/socket/udp_socket_posix.cc
	chromium/net/third_party/quiche/src/quic/core/crypto/transport_parameters.cc
	chromium/net/third_party/quiche/src/quic/core/quic_config.cc
	chromium/net/tools/cert_verify_tool/cert_verify_tool.cc
	chromium/printing/BUILD.gn
	chromium/printing/backend/cups_ipp_constants.h
	chromium/printing/backend/print_backend.h
	chromium/printing/print_settings.cc
	chromium/printing/print_settings.h
	chromium/printing/print_settings_conversion.cc
	chromium/printing/print_settings_conversion_unittest.cc
	chromium/services/device/geolocation/location_arbitrator.cc
	chromium/services/network/network_context.cc
	chromium/services/resource_coordinator/public/cpp/memory_instrumentation/BUILD.gn
	chromium/services/service_manager/embedder/set_process_title.cc -> <deleted>
	chromium/services/service_manager/sandbox/export.h -> <deleted>
	chromium/services/tracing/public/cpp/perfetto/perfetto_traced_process.cc
	chromium/skia/BUILD.gn
	chromium/third_party/angle/gni/angle.gni
	chromium/third_party/angle/src/tests/BUILD.gn
	chromium/third_party/blink/renderer/platform/fonts/font_cache.h
	chromium/third_party/blink/renderer/platform/fonts/skia/font_cache_skia.cc
	chromium/third_party/boringssl/src/include/openssl/rand.h
	chromium/third_party/dav1d/libdav1d/src/x86/cdef.asm -> cdef_avx[2|512].asm
	chromium/third_party/dav1d/libdav1d/src/x86/itx_ssse3.asm
	chromium/third_party/dav1d/libdav1d/src/x86/mc.asm -> mc_avx[2|512].asm
	chromium/third_party/dav1d/libdav1d/src/x86/mc_ssse3.asm -> mc_sse.asm
	chromium/third_party/fontconfig/BUILD.gn
	chromium/third_party/fontconfig/fontconfig.gni
	chromium/third_party/libaom/source/libaom/aom_dsp/x86/aom_convolve_copy_sse2.asm -> aom_convolve_copy_sse2.c
	chromium/third_party/libaom/source/libaom/aom_dsp/x86/quantize_avx_x86_64.asm -> <deleted>
	chromium/third_party/libaom/source/libaom/aom_ports/emms.asm
	chromium/third_party/libaom/source/libaom/aom_ports/x86_abi_support.asm
	chromium/third_party/libaom/source/libaom/third_party/x86inc/x86inc.asm
	chromium/third_party/pdfium/core/fxcrt/cfx_datetime.cpp
	chromium/third_party/pdfium/core/fxcrt/cfx_fileaccess_posix.h
	chromium/third_party/pdfium/core/fxcrt/fx_system.h
	chromium/third_party/pdfium/core/fxge/BUILD.gn
	chromium/third_party/pdfium/core/fxge/fx_ge_linux.cpp
	chromium/third_party/pdfium/third_party/yasm/yasm_assemble.gni
	chromium/third_party/perfetto/include/perfetto/base/build_config.h
	chromium/third_party/perfetto/include/perfetto/base/thread_utils.h
	chromium/third_party/perfetto/src/tracing/core/tracing_service_impl.cc
	chromium/third_party/skia/BUILD.gn
	chromium/third_party/skia/gn/skia.gni
	chromium/third_party/sqlite/src/amalgamation/rename_exports.h -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation/shell/shell.c -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation/sqlite3.c -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation/sqlite3.h -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation_dev/rename_exports.h -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation_dev/shell/shell.c -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation_dev/sqlite3.c -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/amalgamation_dev/sqlite3.h -> <theirs, will reamalgamante>
	chromium/third_party/sqlite/src/configure -> <theirs, will autoreconf>
	chromium/third_party/sqlite/src/configure.ac
	chromium/third_party/sqlite/src/src/os_unix.c
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/netinet/sctp_os_userspace.h
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/netinet/sctp_uio.h
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/netinet/sctp_userspace.c
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/netinet/sctp_usrreq.c
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/user_socket.c
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/user_socketvar.h
	chromium/third_party/yasm/BUILD.gn -> <deleted>
	chromium/third_party/yasm/source/patched-yasm/configure.ac -> <deleted>
	chromium/third_party/yasm/yasm_assemble.gni -> <deleted>
	chromium/ui/aura/window_tree_host_platform.cc
	chromium/ui/base/accelerators/accelerator.cc
	chromium/ui/base/ui_base_export.h -> <deleted>
	chromium/ui/color/BUILD.gn
	chromium/ui/gfx/BUILD.gn
	chromium/ui/gfx/font_render_params.h
	chromium/ui/gfx/native_widget_types.h
	chromium/ui/gl/gl_fence.cc
	chromium/ui/gl/init/BUILD.gn
	chromium/ui/platform_window/platform_window_handler/wm_platform_export.h -> <deleted>
	chromium/ui/shell_dialogs/BUILD.gn
	chromium/url/url_parse_file.cc
	chromium/v8/src/base/once.cc
	chromium/v8/src/base/platform/platform-posix.cc
	chromium/v8/src/snapshot/embedded/platform-embedded-file-writer-generic.cc

+++ chromium/device/fido/cbor_extract.cc
@@ -7,7 +7,7 @@
 #include <type_traits>
 
 #include "base/callback.h"
-#include "base/logging.h"
+#include "base/check_op.h"
 #include "components/cbor/values.h"
 
 namespace device {
@@ -66,8 +66,8 @@ class Extractor {
         step_i_ += key.size() + 1;
         map_it = map.find(cbor::Value(std::move(key)));
       } else {
-        map_it = map.find(
-            cbor::Value(static_cast<int64_t>(key_or_string_indicator)));
+        map_it = map.find(cbor::Value(static_cast<int64_t>(
+            static_cast<int8_t>(key_or_string_indicator))));
       }
 
       const void** output = nullptr;
